chore(prom): rename prometheus(remote storage) to prom-store and promql(HTTP server) to prometheus
