+++ config/standalone.example.toml
@@ -70,12 +70,12 @@ runtime_size = 2
 enable = true
 
 # Prometheus protocol options.
-[prometheus_options]
+[prom_store_options]
 # Whether to enable Prometheus remote write and read in HTTP API, true by default.
 enable = true
 
 # Prom protocol options.
-[prom_options]
+[prometheus_options]
 # Prometheus API server address, "127.0.0.1:4004" by default.
 addr = "127.0.0.1:4004"
