+++ src/servers/src/lib.rs
@@ -31,8 +31,8 @@ pub mod metrics_handler;
 pub mod mysql;
 pub mod opentsdb;
 pub mod postgres;
-pub mod prom;
 pub mod prometheus;
+pub mod prom_store;
 pub mod query_handler;
 pub mod server;
 mod shutdown;
