+++ src/servers/tests/http/mod.rs
@@ -17,4 +17,4 @@ mod http_handler_test;
 mod http_test;
 mod influxdb_test;
 mod opentsdb_test;
-mod prometheus_test;
+mod prom_store_test;
