+++ tests-integration/tests/http.rs
@@ -18,7 +18,7 @@ use common_error::status_code::StatusCode as ErrorCode;
 use serde_json::json;
 use servers::http::handler::HealthResponse;
 use servers::http::{JsonOutput, JsonResponse};
-use servers::prom::{PromJsonResponse, PromResponse};
+use servers::prometheus::{PrometheusJsonResponse, PrometheusResponse};
 use tests_integration::test_util::{
     setup_test_http_app, setup_test_http_app_with_frontend, setup_test_prom_app_with_frontend,
     StorageType,
@@ -328,11 +328,11 @@ pub async fn test_prom_http_api(store_type: StorageType) {
         .send()
         .await;
     assert_eq!(res.status(), StatusCode::OK);
-    let body = serde_json::from_str::<PromJsonResponse>(&res.text().await).unwrap();
+    let body = serde_json::from_str::<PrometheusJsonResponse>(&res.text().await).unwrap();
     assert_eq!(body.status, "success");
     assert_eq!(
         body.data,
-        serde_json::from_value::<PromResponse>(json!(["__name__", "cpu", "host", "memory", "ts"]))
+        serde_json::from_value::<PrometheusResponse>(json!(["__name__", "cpu", "host", "memory", "ts"]))
             .unwrap()
     );
 
@@ -355,11 +355,11 @@ pub async fn test_prom_http_api(store_type: StorageType) {
         .send()
         .await;
     assert_eq!(res.status(), StatusCode::OK);
-    let body = serde_json::from_str::<PromJsonResponse>(&res.text().await).unwrap();
+    let body = serde_json::from_str::<PrometheusJsonResponse>(&res.text().await).unwrap();
     assert_eq!(body.status, "success");
     assert_eq!(
         body.data,
-        serde_json::from_value::<PromResponse>(json!(
+        serde_json::from_value::<PrometheusResponse>(json!(
             [{"__name__" : "demo","ts":"1970-01-01 00:00:00+0000","cpu":"1.1","host":"host1","memory":"2.2"}]
         ))
         .unwrap()
@@ -376,7 +376,7 @@ pub async fn test_prom_http_api(store_type: StorageType) {
     // should return error if there is no match[]
     let res = client.get("/api/v1/label/instance/values").send().await;
     assert_eq!(res.status(), StatusCode::OK);
-    let prom_resp = res.json::<PromJsonResponse>().await;
+    let prom_resp = res.json::<PrometheusJsonResponse>().await;
     assert_eq!(prom_resp.status, "error");
     assert!(prom_resp.error.is_some_and(|err| !err.is_empty()));
     assert!(prom_resp.error_type.is_some_and(|err| !err.is_empty()));
@@ -387,11 +387,11 @@ pub async fn test_prom_http_api(store_type: StorageType) {
         .send()
         .await;
     assert_eq!(res.status(), StatusCode::OK);
-    let body = serde_json::from_str::<PromJsonResponse>(&res.text().await).unwrap();
+    let body = serde_json::from_str::<PrometheusJsonResponse>(&res.text().await).unwrap();
     assert_eq!(body.status, "success");
     assert_eq!(
         body.data,
-        serde_json::from_value::<PromResponse>(json!(["host1", "host2"])).unwrap()
+        serde_json::from_value::<PrometheusResponse>(json!(["host1", "host2"])).unwrap()
     );
 
     // multiple match[]
@@ -400,7 +400,7 @@ pub async fn test_prom_http_api(store_type: StorageType) {
         .send()
         .await;
     assert_eq!(res.status(), StatusCode::OK);
-    let prom_resp = res.json::<PromJsonResponse>().await;
+    let prom_resp = res.json::<PrometheusJsonResponse>().await;
     assert_eq!(prom_resp.status, "success");
     assert!(prom_resp.error.is_none());
     assert!(prom_resp.error_type.is_none());
